# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend Go + Gin untuk aplikasi video conference berbasis LiveKit.\nAuth pakai dua jenis JWT: app JWT (untuk endpoint backend ini) dan LiveKit JWT (untuk connect ke LiveKit server, di-issue oleh /api/token).",
        "title": "Videoconf Backend API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login pakai email + password. Sukses dapet app JWT yang valid 24 jam. Rate-limited (5 req burst per IP).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "login data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "invalid credentials (email/password salah, generic)",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "429": {
                        "description": "rate limit",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Bikin user baru. Email harus unik, password min 8 char, di-hash pakai bcrypt. Sukses langsung dapet app JWT (gak perlu login lagi). Rate-limited (5 req burst per IP).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register user baru",
                "parameters": [
                    {
                        "description": "registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "409": {
                        "description": "email udah dipake",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "429": {
                        "description": "rate limit",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{idOrSlug}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return message DESC by id (paling baru duluan). Pagination cursor-based via `before` (pakai id paling tua dari page sebelumnya).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List chat history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "default 50, max 200",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "kalau diisi, return message dengan id \u003c before",
                        "name": "before",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.messagesListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Persist chat message ke DB. Akses control: owner private OK, anyone untuk public. Real-time delivery di-handle frontend via LiveKit data channel terpisah.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Kirim chat ke room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "message body (max 2000 char)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.sendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Message"
                        }
                    },
                    "400": {
                        "description": "body kosong atau \u003e2000 char",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handlers.authResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/internal_handlers.userPayload"
                }
            }
        },
...
        "internal_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                }
            }
        },
...
        "internal_handlers.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.messagesListResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "body": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "room_id": {
                                "type": "integer"
                            },
                            "sender_id": {
                                "type": "integer"
                            },
                            "sender_name": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "internal_handlers.registerRequest": {
            "type": "object",
            "required": [
                "display_name",
                "email",
                "password"
            ],
            "properties": {
                "display_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "internal_handlers.sendMessageRequest": {
            "type": "object",
            "required": [
                "body"
            ],
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.userPayload": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "videoconf-backend_internal_models.Message": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" lalu spasi, lalu app JWT yang didapat dari /auth/login atau /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}